+++ internal/core/usecase/permission/permission_usecase.go
@@ -1,6 +1,7 @@
-package usecase
+package permission
 
 import (
+	"github.com/br4tech/auth-nex/internal/core/domain"
 	"github.com/br4tech/auth-nex/internal/core/port"
 	"github.com/br4tech/auth-nex/internal/model"
 )
@@ -15,7 +16,18 @@ func NewPermissionUseCase(permissionRepository port.IPermissionRepository) port.
 	}
 }
 
+func (uc *PermissionUseCase) FindRoleByName(name string) (*domain.Role, error) {
+	role, err := uc.permissionRepository.FindRoleByName(name)
+
+	if err != nil {
+		return nil, err
+	}
+
+	return role, nil
+}
+
 func (uc *PermissionUseCase) CreateRole(name string) error {
 	role := &model.Role{Name: name}
+
 	return uc.permissionRepository.CreateRole(role)
 }
