+++ internal/repository/permission_repository.go
@@ -3,6 +3,7 @@ package repository
 import (
 	"github.com/br4tech/auth-nex/internal/core/domain"
 	"github.com/br4tech/auth-nex/internal/core/port"
+	"github.com/br4tech/auth-nex/internal/model"
 	"gorm.io/gorm"
 )
 
@@ -14,10 +15,23 @@ func NewPermissionRepository(db *gorm.DB) port.IPermissionRepository {
 	return &PermissionRepository{db: db}
 }
 
-func (r *PermissionRepository) FindRoleByName(name string) (*domain.User, error) {
-	return nil, nil
+func (r *PermissionRepository) FindRoleByName(name string) (*domain.Role, error) {
+	var roleModel model.Role
+
+	if err := r.db.Where("name=?", name).First(&roleModel).Error; err != nil {
+		return nil, err
+	}
+
+	return roleModel.ToDomain(), nil
 }
 
-func (r *PermissionRepository) CreateRole(role *domain.Role) error {
-	return nil
+func (r *PermissionRepository) CreateRole(role *domain.Role) (*domain.Role, error) {
+	roleModel := new(model.Role)
+	roleModel.FromDomain(role)
+
+	if err := r.db.Create(&roleModel).Error; err != nil {
+		return nil, err
+	}
+
+	return nil, nil
 }
